+++ src/components/Calendar/types.ts
@@ -1,10 +1,11 @@
-import { SchedulerData, SchedulerItemClickData, SchedulerProjectData } from "@/types/global";
+import type { SchedulerData, SchedulerItemClickData, SchedulerProjectData } from "@/types/global";
 
 export type CalendarProps = {
   data: SchedulerData;
   topBarWidth: number;
   onTileClick?: (data: SchedulerProjectData) => void;
   onItemClick?: (data: SchedulerItemClickData) => void;
+  onEmptyCellClick?: (date: Date, id: string) => void;
 };
 
 export type StyledSpanProps = {