+++ src/components/EmptyCellClick/types.ts
@@ -1,0 +1,7 @@
+import type { CellData } from "@/types/global";
+
+export type EmptyCellClickProps = {
+  onEmptyCellClick: (date: Date, id: string) => void;
+  cellData: CellData;
+  zoom: number;
+};