+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [core] custom data tile render ([#6](https://github.com/CoderKaine/react-scheduler/pull/6)) [@lau-sam](https://github.com/lau-sam)
 - [docs] split README.md into multiples docs files ([#7](https://github.com/CoderKaine/react-scheduler/pull/7)) [@lau-sam](https://github.com/lau-sam)
 - [LeftColumnItem] default fallback for broken icon links ([#9](https://github.com/CoderKaine/react-scheduler/pull/9)) [@lau-sam](https://github.com/lau-sam)
+- [core] Allow click action on a cell that has no data ([#11](https://github.com/CoderKaine/react-scheduler/pull/11)) [@lau-sam](https://github.com/lau-sam)
 
 ## 0.1.5
 Forked from [@bitnoi.se/react-scheduler](https://github.com/Bitnoise/react-scheduler)